+++ polylogue/drive_client.py
@@ -71,8 +71,10 @@ class DriveClient:
     def _prompt_for_credentials(self) -> Path:
         self.ui.banner("Google Drive access needs credentials", "Download OAuth client for Desktop app")
         clipboard_checked = False
+        # Only auto-read clipboard if explicitly enabled (security: don't read clipboard without permission)
+        allow_clipboard = os.environ.get("POLYLOGUE_ALLOW_CLIPBOARD", "").lower() in ("1", "true", "yes")
         while True:
-            if not clipboard_checked:
+            if not clipboard_checked and allow_clipboard:
                 clipboard_checked = True
                 clip_path = self._try_clipboard_credentials()
                 if clip_path:
