+++ polylogue/ui/facade.py
@@ -1,12 +1,18 @@
+"""Pure Python UI facade using Questionary and Rich (no external binaries)."""
 from __future__ import annotations
 
-import shutil
-import subprocess
+import difflib
 from dataclasses import dataclass, field
 from typing import Iterable, List, Optional, Protocol
 
+import questionary
+from pygments import highlight
+from pygments.formatters import TerminalFormatter
+from pygments.lexers import get_lexer_by_name
 from rich.console import Console
+from rich.markdown import Markdown
 from rich.panel import Panel
+from rich.syntax import Syntax
 from rich.text import Text
 
 
@@ -28,136 +34,158 @@ class PlainConsole:
 
 @dataclass
 class ConsoleFacade:
-    """Base facade for interacting with terminal UI."""
+    """Pure Python facade for terminal UI - no external binaries required."""
 
     plain: bool
     console: ConsoleLike = field(init=False)
 
+    def __post_init__(self) -> None:
+        if self.plain:
+            self.console = PlainConsole()
+        else:
+            self.console = Console(no_color=False, force_terminal=True)
+
     def banner(self, title: str, subtitle: Optional[str] = None) -> None:
+        """Display a banner message."""
         if self.plain:
             self.console.print(f"== {title} ==")
             if subtitle:
                 self.console.print(subtitle)
             return
-        self._interactive_banner(title, subtitle)
+
+        body = title if not subtitle else f"{title}\n{subtitle}"
+        panel = Panel(Text(body, style="bold cyan"), border_style="blue")
+        self.console.print(panel)
 
     def summary(self, title: str, lines: Iterable[str]) -> None:
+        """Display a summary panel."""
         text = "\n".join(lines)
         if self.plain:
             self.console.print(f"-- {title} --")
             if text:
                 self.console.print(text)
             return
-        self._interactive_summary(title, text)
+
+        panel = Panel(Text(text), title=title, title_align="left")
+        self.console.print(panel)
 
     def confirm(self, prompt: str, *, default: bool = True) -> bool:
+        """Ask for confirmation."""
         if self.plain:
             return default
-        return self._interactive_confirm(prompt, default=default)
+        return questionary.confirm(prompt, default=default).ask() or default
 
     def choose(self, prompt: str, options: List[str]) -> Optional[str]:
+        """Choose from a list of options."""
         if not options:
             return None
         if self.plain:
             return None
-        return self._interactive_choose(prompt, options)
+        result = questionary.select(prompt, choices=options).ask()
+        return result
 
     def input(self, prompt: str, *, default: Optional[str] = None) -> Optional[str]:
+        """Get text input from user."""
         if self.plain:
             return default
-        return self._interactive_input(prompt, default=default)
+        result = questionary.text(prompt, default=default or "").ask()
+        return result if result else default
 
-    # Interactive implementations -----------------------------------------
-    def _interactive_banner(self, title: str, subtitle: Optional[str]) -> None:
-        body = title if not subtitle else f"{title}\n{subtitle}"
-        result = subprocess.run(
-            [
-                "gum",
-                "style",
-                "--border",
-                "rounded",
-                "--margin",
-                "1",
-                "--padding",
-                "1",
-                body,
-            ],
-            check=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-        )
-        if result.stdout.strip():
-            self.console.print(result.stdout.rstrip(), markup=False, highlight=False)
-        else:
-            self.console.print(body, markup=False, highlight=False)
-
-    def _interactive_summary(self, title: str, text: str) -> None:
-        result = subprocess.run(
-            ["gum", "format"],
-            input=f"## {title}\n{text}\n",
-            text=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
-        output = result.stdout.rstrip()
-        if output:
-            self.console.print(output, markup=False, highlight=False)
+    def render_markdown(self, content: str) -> None:
+        """Render Markdown content."""
+        if self.plain:
+            self.console.print(content)
             return
-        self.console.print(Panel(Text(text), title=title))
 
-    def _interactive_confirm(self, prompt: str, *, default: bool) -> bool:
-        cmd: List[str] = ["gum", "confirm", "--prompt", prompt]
-        if default:
-            cmd.append("--default")
-        result = subprocess.run(cmd, check=False)
-        return result.returncode == 0
+        md = Markdown(content)
+        self.console.print(md)
 
-    def _interactive_choose(self, prompt: str, options: List[str]) -> Optional[str]:
-        result = subprocess.run(
-            ["gum", "choose", "--header", prompt, *options],
-            check=False,
-            stdout=subprocess.PIPE,
-            text=True,
+    def render_code(self, code: str, language: str = "python") -> None:
+        """Render syntax-highlighted code."""
+        if self.plain:
+            self.console.print(code)
+            return
+
+        syntax = Syntax(code, language, theme="monokai", line_numbers=True)
+        self.console.print(syntax)
+
+    def render_diff(self, old_text: str, new_text: str, filename: str = "file") -> None:
+        """Render a diff between two texts."""
+        old_lines = old_text.splitlines(keepends=True)
+        new_lines = new_text.splitlines(keepends=True)
+
+        diff = difflib.unified_diff(
+            old_lines,
+            new_lines,
+            fromfile=f"a/{filename}",
+            tofile=f"b/{filename}",
+            lineterm=""
         )
-        choice = result.stdout.strip()
-        return choice or None
 
-    def _interactive_input(self, prompt: str, *, default: Optional[str]) -> Optional[str]:
-        cmd: List[str] = ["gum", "input", "--placeholder", prompt]
-        if default:
-            cmd.extend(["--value", default])
-        result = subprocess.run(cmd, check=False, stdout=subprocess.PIPE, text=True)
-        value = result.stdout.strip()
-        if value:
-            return value
-        return default
+        diff_text = "".join(diff)
+
+        if self.plain:
+            self.console.print(diff_text)
+            return
+
+        # Use pygments for diff highlighting
+        try:
+            lexer = get_lexer_by_name("diff")
+            highlighted = highlight(diff_text, lexer, TerminalFormatter())
+            self.console.print(highlighted, markup=False, highlight=False)
+        except Exception:
+            # Fallback to plain diff
+            self.console.print(diff_text, markup=False)
+
+    def error(self, message: str) -> None:
+        """Display an error message."""
+        if self.plain:
+            self.console.print(f"ERROR: {message}")
+        else:
+            self.console.print(f"[bold red]ERROR:[/bold red] {message}")
+
+    def warning(self, message: str) -> None:
+        """Display a warning message."""
+        if self.plain:
+            self.console.print(f"WARNING: {message}")
+        else:
+            self.console.print(f"[bold yellow]WARNING:[/bold yellow] {message}")
+
+    def success(self, message: str) -> None:
+        """Display a success message."""
+        if self.plain:
+            self.console.print(f"SUCCESS: {message}")
+        else:
+            self.console.print(f"[bold green]✓[/bold green] {message}")
+
+    def info(self, message: str) -> None:
+        """Display an info message."""
+        if self.plain:
+            self.console.print(message)
+        else:
+            self.console.print(f"[cyan]ℹ[/cyan] {message}")
 
 
 @dataclass
 class PlainConsoleFacade(ConsoleFacade):
+    """Plain console facade for non-interactive environments."""
+
     def __post_init__(self) -> None:
-        self.console = PlainConsole()
         self.plain = True
+        self.console = PlainConsole()
 
 
 @dataclass
 class InteractiveConsoleFacade(ConsoleFacade):
-    """Rich + gum backed facade that requires the interactive toolchain."""
+    """Interactive console facade using pure Python libraries."""
 
     def __post_init__(self) -> None:
-        missing = [cmd for cmd in ("gum", "sk", "bat", "glow", "delta") if shutil.which(cmd) is None]
-        if missing:
-            missing_cmds = ", ".join(missing)
-            raise RuntimeError(
-                f"Interactive dependencies missing: {missing_cmds}. "
-                "Enter the Polylogue devshell (`nix develop`) or install the required tools."
-            )
+        self.plain = False
         self.console = Console(no_color=False, force_terminal=True)
 
 
 def create_console_facade(plain: bool) -> ConsoleFacade:
+    """Create a console facade with no external binary dependencies."""
     if plain:
         return PlainConsoleFacade(plain=True)
     return InteractiveConsoleFacade(plain=False)
